demux: mkv: use size_t for levels in m_el table
It should never be negative.

+++ modules/demux/mkv/Ebml_parser.hpp
@@ -56,12 +56,12 @@ class EbmlParser
 
     demux_t     *p_demux;
     matroska_iostream_c  *m_es;
-    int          mi_level;
+    size_t       mi_level;
     EbmlElement *m_el[M_EL_MAXSIZE];
 
     EbmlElement *m_got;
 
-    int          mi_user_level;
+    size_t       mi_user_level;
     bool         mb_keep;
     /* Allow dummy/unknown EBML elements */
     bool         mb_dummy;
